requirements.txt

beautifulsoup4==4.8.1

Brotli==1.0.7
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cloudscraper==1.2.20
cryptography==2.8
dnspython==1.16.0
future==0.18.2
idna==2.8

pycparser==2.19
pymongo==3.10.1
PySocks==1.7.1
python-telegram-bot==12.2.0

redis==3.3.11
requests==2.22.0
requests-toolbelt==0.9.1
schedule==0.6.0
six==1.13.0
soupsieve==1.9.5
tornado==6.0.3
urllib3==1.25.7

